+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added SOPS support to ansible.cfg file, as provided through helm chart
-
+- Added CFS clean-up of IMS jobs that it creates for image customization.
 
 ## [1.19.0] - 8/18/2023
 ### Changed